// src/common/enums/prediction-type.enum.ts
export enum PredictionType {
  MATCH_RESULT = "MATCH_RESULT",
  OVER_UNDER = "OVER_UNDER",
  BOTH_TEAMS_TO_SCORE = "BOTH_TEAMS_TO_SCORE",
  DOUBLE_CHANCE = "DOUBLE_CHANCE",
  HANDICAP = "HANDICAP",
  CORRECT_SCORE = "CORRECT_SCORE",
  FIRST_GOAL_SCORER = "FIRST_GOAL_SCORER",
  ANY_OTHER = "ANY_OTHER",
}